CHG: Replaced logger with new 'gnklog'
This also triggered quite a lot of new warnings due to C++ packed parameters instead of C variadic function args. Mostly enum -> int (for class enum's) but also a few 'char32_t' which can't be mixed with regular char.

+++ src/Core/Controllers/QuickCommandController.cpp
@@ -71,13 +71,13 @@ bool QuickCommandController::HandleActionInQuickCmdState(const KeyPressAction &k
             return true;
         case kAction::kActionStartSearch :
             Editor::Instance().GetActiveModel()->ResetSearchHitIndex();
-            logger->Debug("Entering search, key: %c",kpAction.keyPress.key);
+            logger->Debug("Entering search, key: %c",(int)kpAction.keyPress.key);
             ChangeState(State::SearchState);
             cmdInputBaseController.DefaultEditLine(cursor, cmdInput, kpAction.keyPress, true);
             break;
         case kAction::kActionLastSearch :
             Editor::Instance().GetActiveModel()->ResetSearchHitIndex();
-            logger->Debug("Entering search, key: %c",kpAction.keyPress.key);
+            logger->Debug("Entering search, key: %c",(int)kpAction.keyPress.key);
             cmdInputBaseController.DefaultEditLine(cursor, cmdInput, kpAction.keyPress, true);
             if (!searchHistory.empty()) {
                 auto &lastSearchItem = searchHistory.back();
